# interp: increase speed of findSegment